Revert "ion: add COMPAT_ION_IOC_SUNXI_FLUSH_RANGE in compat_ion_ioctl for support 32bit app."
This reverts commit 582e61c0b38506c69c235da9cb8adb5214c7ae17.

+++ drivers/staging/android/uapi/ion_sunxi.h
@@ -24,8 +24,8 @@ typedef struct {
 
 typedef struct {
 	ion_user_handle_t handle;
-	unsigned int phys_addr;
-	unsigned int size;
+	unsigned long phys_addr;
+	unsigned long size;
 }sunxi_phys_data;
 
 #define ION_IOC_SUNXI_FLUSH_RANGE           5
